Add some continuous display mode stuff...
Added some simple mode where it reads picture and displays it and does it again.

It is doing this by using frame buffer for display and once you have a frame, from camera it then waits if necessary for the DMA update to complete from the camera and then starts up displaying the next frame.